TX: Use correct integer type for bundle dwell time
uint8_t is much too small, we always use uint64_t when dealing with
bundle age values, as we should do here.

Found by Clang's UB sanitizer.

+++ components/cla/cla_contact_tx_task.c
@@ -34,7 +34,7 @@ static void prepare_bundle_for_forwarding(struct bundle *bundle)
 		blocks = &(*blocks)->next;
 	}
 
-	const uint8_t dwell_time_ms = hal_time_get_timestamp_ms() -
+	const uint64_t dwell_time_ms = hal_time_get_timestamp_ms() -
 		bundle->reception_timestamp_ms;
 
 	// BPv7 5.4-4: "If the bundle has a bundle age block ... at the last
